ENH: Created new python files for supporting batch mode registration (issue #119) - also added small widget SliceTrackerRegistration which has required selectors and an execute button - execution from cmd with: ./Slicer --no-splash --python-script ../SliceTrackerRegistration.py  -fv fixedVolume -mv movingVolume -fl fixedLabel -ml movingLabel -o outputDirectory - moved constants to separate python file - moved RegistrationResults and RegistrationResult class to RegistrationData.py -- those classes are used by both SliceTracker and SliceTrackerRegistration
