+++ SliceTrackerUtils/Constants.py
@@ -1,0 +1,53 @@
+import qt
+import slicer
+
+
+class DICOMTAGS:
+
+  PATIENT_NAME          = '0010,0010'
+  PATIENT_ID            = '0010,0020'
+  PATIENT_BIRTH_DATE    = '0010,0030'
+  SERIES_DESCRIPTION    = '0008,103E'
+  SERIES_NUMBER         = '0020,0011'
+  STUDY_DATE            = '0008,0020'
+  STUDY_TIME            = '0008,0030'
+  ACQUISITION_TIME      = '0008,0032'
+
+
+class COLOR:
+
+  RED = qt.QColor(qt.Qt.red)
+  YELLOW = qt.QColor(qt.Qt.yellow)
+  GREEN = qt.QColor(qt.Qt.darkGreen)
+  GRAY = qt.QColor(qt.Qt.gray)
+
+
+class STYLE:
+
+  WHITE_BACKGROUND            = 'background-color: rgb(255,255,255)'
+  LIGHT_GRAY_BACKGROUND       = 'background-color: rgb(230,230,230)'
+  ORANGE_BACKGROUND           = 'background-color: rgb(255,102,0)'
+  YELLOW_BACKGROUND           = 'background-color: yellow;'
+  GREEN_BACKGROUND            = 'background-color: green;'
+  GRAY_BACKGROUND             = 'background-color: gray;'
+  RED_BACKGROUND              = 'background-color: red;'
+
+
+class SliceTrackerConstants(object):
+
+  LEFT_VIEWER_SLICE_ANNOTATION_TEXT = 'BIOPSY PLAN'
+  RIGHT_VIEWER_SLICE_ANNOTATION_TEXT = 'TRACKED TARGETS'
+  RIGHT_VIEWER_SLICE_TRANSFORMED_ANNOTATION_TEXT = 'OLD'
+  RIGHT_VIEWER_SLICE_NEEDLE_IMAGE_ANNOTATION_TEXT = 'NEW'
+  APPROVED_RESULT_TEXT_ANNOTATION = "approved"
+  REJECTED_RESULT_TEXT_ANNOTATION = "rejected"
+  SKIPPED_RESULT_TEXT_ANNOTATION = "skipped"
+
+  LAYOUT_RED_SLICE_ONLY = slicer.vtkMRMLLayoutNode.SlicerLayoutOneUpRedSliceView
+  LAYOUT_FOUR_UP = slicer.vtkMRMLLayoutNode.SlicerLayoutFourUpView
+  LAYOUT_SIDE_BY_SIDE = slicer.vtkMRMLLayoutNode.SlicerLayoutSideBySideView
+  ALLOWED_LAYOUTS = [LAYOUT_SIDE_BY_SIDE, LAYOUT_FOUR_UP]
+
+  COVER_PROSTATE = "COVER PROSTATE"
+  COVER_TEMPLATE = "COVER TEMPLATE"
+  GUIDANCE_IMAGE = "GUIDANCE"
